Merge primary repository into fork

+++ tests/utils.py
@@ -36,7 +36,7 @@ def capture_output(redirect_stderr=True):
         sys.stdout, sys.stderr = oldout, olderr
 
 
-def adjust(text):
+def adjust(text, run_in_function=False):
     """Adjust a code sample to remove leading whitespace."""
     lines = text.split('\n')
     if len(lines) == 1:
@@ -47,14 +47,27 @@ def adjust(text):
     first_line = lines[0].lstrip()
     n_spaces = len(lines[0]) - len(first_line)
 
-    return '\n'.join(line[n_spaces:] for line in lines)
+    if run_in_function:
+        n_spaces = n_spaces - 4
+
+    final_lines = [line[n_spaces:] for line in lines]
+
+    if run_in_function:
+        final_lines = [
+            "def test_function():",
+        ] + final_lines + [
+            "test_function()",
+            "print('Function done.')"
+        ]
+
+    return '\n'.join(final_lines)
 
 
-def runAsPython(test_dir, main_code, extra_code=None):
+def runAsPython(test_dir, main_code, extra_code=None, run_in_function=False, args=None):
     """Run a block of Python code with the Python interpreter."""
     # Output source code into test directory
     with open(os.path.join(test_dir, 'test.py'), 'w') as py_source:
-        py_source.write(main_code)
+        py_source.write(adjust(main_code, run_in_function=run_in_function))
 
     if extra_code:
         for name, code in extra_code.items():
@@ -68,8 +81,11 @@ def runAsPython(test_dir, main_code, extra_code=None):
             with open(os.path.join(test_dir, *path), 'w') as py_source:
                 py_source.write(adjust(code))
 
+    if args is None:
+        args = []
+
     proc = subprocess.Popen(
-        ["python", "test.py"],
+        ["python", "test.py"] + args,
         stdin=subprocess.PIPE,
         stdout=subprocess.PIPE,
         stderr=subprocess.STDOUT,
@@ -80,7 +96,7 @@ def runAsPython(test_dir, main_code, extra_code=None):
     return out[0].decode('utf8')
 
 
-def runAsJava(test_dir, main_code, extra_code=None):
+def runAsJava(test_dir, main_code, extra_code=None, run_in_function=False, args=None):
     """Run a block of Python code as a Java program."""
     # Output source code into test directory
     transpiler = Transpiler()
@@ -88,16 +104,19 @@ def runAsJava(test_dir, main_code, extra_code=None):
     # Don't redirect stderr; we want to see any errors from the transpiler
     # as top level test failures.
     with capture_output(redirect_stderr=False):
-        transpiler.transpile_string("test.py", main_code)
+        transpiler.transpile_string("test.py", adjust(main_code, run_in_function=run_in_function))
 
         if extra_code:
             for name, code in extra_code.items():
-                transpiler.transpile_string("%s.py" % name, code)
+                transpiler.transpile_string("%s.py" % name, adjust(code))
 
     transpiler.write(test_dir, verbosity=0)
 
+    if args is None:
+        args = []
+
     proc = subprocess.Popen(
-        ["java", "-classpath", "../../python.jar:.", "-XX:-UseSplitVerifier", "python.test"],
+        ["java", "-classpath", "../../dist/python-java.jar:.", "-XX:-UseSplitVerifier", "python.test"] + args,
         stdin=subprocess.PIPE,
         stdout=subprocess.PIPE,
         stderr=subprocess.STDOUT,
@@ -108,21 +127,56 @@ def runAsJava(test_dir, main_code, extra_code=None):
     return out[0].decode('utf8')
 
 
-JAVA_EXCEPTION = re.compile('^Exception in thread "main" ([^\n]+\nCaused by: )?org\.python\.exceptions\.([\w]+): ([^\n]+)\n([^\n]+\n)+', re.MULTILINE)
+JAVA_EXCEPTION = re.compile(
+    '(org\.python\.exceptions\.([\w]+): ([^\n]+)\n' +
+    '(\s+at .+\((((.*):(\d+))|(Native Method))\)\n)+)?' +
+    '((Exception in thread "\w+" org\.python\.exceptions\.(?P<exception1>[\w]+): (?P<message1>[^\n]+))|' +
+    '(Exception in thread "\w+" [^\n]+\n' +
+    'Caused by: org\.python\.exceptions\.(?P<exception2>[\w]+): (?P<message2>[^\n]+)))\n' +
+    '(?P<trace>(\s+at .+\((((.*):(\d+))|(Native Method))\)\n)+)'
+)
+JAVA_STACK = re.compile('\s+at (?P<module>.+)\((((?P<file>.*):(?P<line>\d+))|(Native Method))\)')
 JAVA_FLOAT = re.compile('(\d+)E(-)?(\d+)')
 
-PYTHON_EXCEPTION = re.compile('^Traceback \(most recent call last\):\n(  .*\n)+(.*): (.*)\n', re.MULTILINE)
+# PYTHON_EXCEPTION = re.compile('Traceback \(most recent call last\):\n(  File ".*", line \d+, in .*\n)(    .*\n  File "(?P<file>.*)", line (?P<line>\d+), in .*\n)+(?P<exception>.*): (?P<message>.*\n)')
+PYTHON_EXCEPTION = re.compile('Traceback \(most recent call last\):\n(  File "(?P<file>.*)", line (?P<line>\d+), in .*\n    .*\n)+(?P<exception>.*?): (?P<message>.*\n)')
+PYTHON_STACK = re.compile('  File "(?P<file>.*)", line (?P<line>\d+), in .*\n    .*\n')
 PYTHON_FLOAT = re.compile('(\d+)e(-)?0?(\d+)')
 
 
 def cleanse_java(input):
-    out = JAVA_EXCEPTION.sub('### EXCEPTION ###\n\\2: \\3\n', input)
-    return JAVA_FLOAT.sub('\\1e\\2\\3', out)
+    try:
+        out = JAVA_EXCEPTION.sub('### EXCEPTION ###\n\\g<exception2>: \\g<message2>\n\\g<trace>', input)
+    except:
+        out = JAVA_EXCEPTION.sub('### EXCEPTION ###\n\\g<exception1>: \\g<message1>\n\\g<trace>', input)
+    stack = JAVA_STACK.findall(out)
+    out = JAVA_STACK.sub('', out)
+    out = '%s%s%s' % (
+        out,
+        '\n'.join([
+            "    %s:%s" % (s[3], s[4])
+            for s in stack[::-1]
+            if s[0].startswith('python.')
+        ]),
+        '\n' if stack else ''
+    )
+    return JAVA_FLOAT.sub('\\1e\\2\\3', out).replace("'python.test'", '***EXECUTABLE***')
 
 
 def cleanse_python(input):
-    out = PYTHON_EXCEPTION.sub('### EXCEPTION ###\n\\2: \\3\n', input)
-    return PYTHON_FLOAT.sub('\\1e\\2\\3', out)
+    out = PYTHON_EXCEPTION.sub('### EXCEPTION ###\n\\g<exception>: \\g<message>', input)
+    stack = PYTHON_STACK.findall(input)
+    out = '%s%s%s' % (
+        out,
+        '\n'.join(
+            [
+                "    %s:%s" % (s[0], s[1])
+                for s in stack
+            ]
+        ),
+        '\n' if stack else ''
+    )
+    return PYTHON_FLOAT.sub('\\1e\\2\\3', out).replace("'test.py'", '***EXECUTABLE***')
 
 
 class TranspileTestCase(TestCase):
@@ -159,35 +213,66 @@ class TranspileTestCase(TestCase):
         java = adjust(java)
         self.assertEqual(debug.getvalue(), java[1:])
 
-    def assertCodeExecution(self, code, message=None, extra_code=None):
+    def assertCodeExecution(self, code, message=None, extra_code=None, run_in_global=True, run_in_function=True, args=None):
         "Run code as native python, and under Java and check the output is identical"
         self.maxDiff = None
-        code = adjust(code)
-
-        try:
-            # Create the temp directory into which code will be placed
-            test_dir = os.path.join(os.path.dirname(__file__), 'temp')
+        #==================================================
+        # Pass 1 - run the code in the global context
+        #==================================================
+        if run_in_global:
             try:
-                os.mkdir(test_dir)
-            except FileExistsError:
-                pass
+                # Create the temp directory into which code will be placed
+                test_dir = os.path.join(os.path.dirname(__file__), 'temp')
+                try:
+                    os.mkdir(test_dir)
+                except FileExistsError:
+                    pass
 
-            # Run the code as Python and as Java.
-            py_out = runAsPython(test_dir, code, extra_code)
-            java_out = runAsJava(test_dir, code, extra_code)
-        except Exception as e:
-            self.fail(e)
-        finally:
-            # Clean up the test directory where the class file was written.
-            shutil.rmtree(test_dir)
+                # Run the code as Python and as Java.
+                py_out = runAsPython(test_dir, code, extra_code, False, args=args)
+                java_out = runAsJava(test_dir, code, extra_code, False, args=args)
+            except Exception as e:
+                self.fail(e)
+            finally:
+                # Clean up the test directory where the class file was written.
+                shutil.rmtree(test_dir)
 
-        # Cleanse the Python and Java output, producing a simple
-        # normalized format for exceptions, floats etc.
-        java_out = cleanse_java(java_out)
-        py_out = cleanse_python(py_out)
+            # Cleanse the Python and Java output, producing a simple
+            # normalized format for exceptions, floats etc.
+            java_out = cleanse_java(java_out)
+            py_out = cleanse_python(py_out)
 
-        # Confirm that the output of the Java code is the same as the Python code.
-        self.assertEqual(java_out, py_out, message)
+            # Confirm that the output of the Java code is the same as the Python code.
+            self.assertEqual(java_out, py_out, 'Global context: %s' % message)
+
+        #==================================================
+        # Pass 2 - run the code in a function's context
+        #==================================================
+        if run_in_function:
+            try:
+                # Create the temp directory into which code will be placed
+                test_dir = os.path.join(os.path.dirname(__file__), 'temp')
+                try:
+                    os.mkdir(test_dir)
+                except FileExistsError:
+                    pass
+
+                # Run the code as Python and as Java.
+                py_out = runAsPython(test_dir, code, extra_code, True, args=args)
+                java_out = runAsJava(test_dir, code, extra_code, True, args=args)
+            except Exception as e:
+                self.fail(e)
+            finally:
+                # Clean up the test directory where the class file was written.
+                shutil.rmtree(test_dir)
+
+            # Cleanse the Python and Java output, producing a simple
+            # normalized format for exceptions, floats etc.
+            java_out = cleanse_java(java_out)
+            py_out = cleanse_python(py_out)
+
+            # Confirm that the output of the Java code is the same as the Python code.
+            self.assertEqual(java_out, py_out, 'Function context: %s' % message)
 
 
 def _unary_test(test_name, operation):
@@ -278,6 +363,13 @@ class BinaryOperationTestCase:
         vars()['test_and_%s' % datatype] = _binary_test('test_and_%s' % datatype, 'x & y', examples)
         vars()['test_xor_%s' % datatype] = _binary_test('test_xor_%s' % datatype, 'x ^ y', examples)
         vars()['test_or_%s' % datatype] = _binary_test('test_or_%s' % datatype, 'x | y', examples)
+
+        vars()['test_lt_%s' % datatype] = _binary_test('test_lt_%s' % datatype, 'x < y', examples)
+        vars()['test_le_%s' % datatype] = _binary_test('test_le_%s' % datatype, 'x <= y', examples)
+        vars()['test_gt_%s' % datatype] = _binary_test('test_gt_%s' % datatype, 'x > y', examples)
+        vars()['test_ge_%s' % datatype] = _binary_test('test_ge_%s' % datatype, 'x >= y', examples)
+        vars()['test_eq_%s' % datatype] = _binary_test('test_eq_%s' % datatype, 'x == y', examples)
+        vars()['test_ne_%s' % datatype] = _binary_test('test_ne_%s' % datatype, 'x != y', examples)
 
 
 def _inplace_test(test_name, operation, examples):
